backend/requirements.txt
fastapi
uvicorn
pydantic
boto3
langgraph
langchain-aws
langchain-core
langchain-text-splitters
pytest
python-dotenv
requests
beautifulsoup4
pinecone
pypdf
duckdb
openpyxl
pandas
python-multipart
structlog
asgi-correlation-id
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-botocore
opentelemetry-exporter-otlp
opentelemetry-sdk-extension-aws
psycopg2-binary
presidio-analyzer
presidio-anonymizer
spacy
